ivp_utility: x86_64 Fix compiler warnings about type truncation

+++ ivp_utility/ivu_vhash.hxx
@@ -40,7 +40,7 @@ protected:
     IVP_VHash(IVP_VHash_Elem *static_elems, int size);// assert(size = 2**x)
     virtual IVP_BOOL compare(void *elem0, void *elem1)const = 0;  // return TRUE if equal
 public:
-    static inline int hash_index(const char *data, int size);            // useable index calculation, result is [0,0xfffffff]
+    static inline int hash_index(const char *data, intp size);            // useable index calculation, result is [0,0xfffffff]
     static inline int fast_hash_index(int key);       // useable index calculation when size == 4 , result is [0,0xfffffff]
     static inline int fast_hash_index(long long key);       // useable index calculation when size == 4 , result is [0,0xfffffff]
 
@@ -77,10 +77,10 @@ public:
 
 
 // basic function for calculating the hash_index
-inline int IVP_VHash::hash_index(const char *key, int key_size){
+inline int IVP_VHash::hash_index(const char *key, intp key_size){
 	unsigned int c;		
 	unsigned int index = 0xffffffffL;
-	int i;
+	intp i;
 	for (i=key_size-1;i>=0;i--){
 	    c = *((unsigned char *)(key++));
 	    index = IVP_Hash_crctab[((int) index ^ c) & 0xff] ^ (index >> 8);
